+++ insh-db/migrations/2026-09-08-000000_create_dir_history/down.sql
@@ -1,0 +1,1 @@
+DROP TABLE dir_history;