+++ insh/src/components/common/mod.rs
@@ -1,7 +1,7 @@
 //! Common components.
 
 mod dir;
-pub use dir::{Dir, Event as DirEvent, Props as DirProps};
+pub use dir::{Dir, Effect as DirEffect, Event as DirEvent, Props as DirProps};
 
 mod footer;
 pub use footer::{Footer, Info as FooterInfo, Props as FooterProps};