+++ inshd/src/request_handlers/mod.rs
@@ -6,8 +6,10 @@ mod get_file_contents;
 mod get_files;
 mod search_phrase;
 mod stream_logs;
+mod suggest_dir;
 mod suggest_find_pattern;
 mod suggest_search_phrase;
+mod visit_dir;
 
 pub use create_file::CreateFile;
 pub use database_info::DatabaseInfo;
@@ -16,5 +18,7 @@ pub use get_file_contents::GetFileContents;
 pub use get_files::GetFiles;
 pub use search_phrase::SearchPhrase;
 pub use stream_logs::StreamLogs;
+pub use suggest_dir::SuggestDir;
 pub use suggest_find_pattern::SuggestFindPattern;
 pub use suggest_search_phrase::SuggestSearchPhrase;
+pub use visit_dir::VisitDir;